% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanNA}
\alias{cleanNA}
\title{Remove the rows with NA of the SummarizeExperiment.}
\usage{
cleanNA(sumex, Ncolumns = 3)
}
\arguments{
\item{sumex}{is the SummarizeExperiment}

\item{Ncolumns}{The number of identifiers of the SummarizedExperiment.}
}
\value{
SummarizedExperiment
}
\description{
Removes rows that have as many NAs in the identifiers as the specified value by argument.
The number of identifiers of the SummarizedExperiment should be introduced as an argument,
in this way we can remove the rows that only have NA values in the different identifiers.
}
\seealso{
Other NA clean: 
\code{\link{unifyeset}()}
}
\concept{NA clean}
