% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ensembl2url}
\alias{ensembl2url}
\title{URL's from Ensembl identifiers}
\usage{
ensembl2url(id)
}
\arguments{
\item{id}{Ensembl identifiers}
}
\value{
The corresponding URL's in the database
}
\description{
It makes the url's from Ensembl identifiers
}
\seealso{
Other URL generation: 
\code{\link{entrezid2url}()},
\code{\link{kegg2url}()}
}
\concept{URL generation}
