% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eset2df}
\alias{eset2df}
\title{Obtain a data.frame from expression set with p.values using rowttests or limma.}
\usage{
eset2df(eset, method = 1)
}
\arguments{
\item{eset}{Is the ExpressionSet object}

\item{method}{1 To determinate the p-values with genefilter library, and 2 to determinate the p-values with limma.}
}
\value{
data.frame
}
\description{
Generates a dataframe with the fData information of the expression set
and adds the p-value and adjusted p-value information using rowttests of
genefilter or limma library.
}
\concept{data.frame generation}
