% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneSet_Analysis_GSE115484}
\alias{GeneSet_Analysis_GSE115484}
\title{GeneSet_Analysis_GSE115484}
\format{
analysis
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE115484}
 
 \url{https://pubmed.ncbi.nlm.nih.gov/31234802/}
}
\description{
Sus scrofa

Expression data from Microarray experiment in muscle and liver tissues.

Expression profiling by array.

Gene set analysis with the data generated in GSE115484 using EnrichmentBrowser package (tarea51.Rmd) and GeneSetTest function from tami (tarea52.Rmd).
}
\keyword{datasets}
