% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE115484}
\alias{GSE115484}
\title{GSE115484}
\format{
SummarizedExperiment
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE115484}
 
 \url{https://pubmed.ncbi.nlm.nih.gov/31234802/}
}
\description{
Sus scrofa

Expression data from Microarray experiment in muscle and liver tissues

Expression profiling by array.

In the current work we have compared the transcriptomic and eQTL profiles of the porcine skeletal muscle and
liver by using a data set of 103 Duroc pigs genotyped with the Illumina SNP60 BeadChip and with available 
microarray measurements of gene expression for both tissues.'
}
\examples{
data(gse115484, package="asierortega")
}
\keyword{datasets}
