% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE115484_DE_analysis}
\alias{GSE115484_DE_analysis}
\title{GSE115484_DE_analysis}
\format{
analysis
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE115484}
 
 \url{https://pubmed.ncbi.nlm.nih.gov/31234802/}
}
\description{
Sus scrofa

Expression data from Microarray experiment in muscle and liver tissues.

Expression profiling by array.

Differential expression analysis with the data generated in GSE115484 using genefilter and limma methods, tarea3.Rmd.
}
\keyword{datasets}
