% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kegg2url}
\alias{kegg2url}
\title{URL's from KEGG pathway identifiers}
\usage{
kegg2url(id)
}
\arguments{
\item{id}{KEGG pathway identifiers}
}
\value{
The corresponding URL's in the database
}
\description{
It makes the url's from KEGG pathway identifiers
}
\seealso{
Other URL generation: 
\code{\link{ensembl2url}()},
\code{\link{entrezid2url}()}
}
\concept{URL generation}
