% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeReport}
\alias{makeReport}
\title{Generates an HTML report of a given data.frame.}
\usage{
makeInform(df, name, path=".")
}
\arguments{
\item{df}{Is the dataframe}

\item{name}{Name of the inform/file}

\item{path}{Is the path where you want to generate the HTML inform. As default value will be current working directory.}
}
\value{
Generates an HTML report.
}
\description{
Generates an HTML report of a given given data.frame in a exact path given as argument.
}
\concept{report}
