% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{overex2df}
\alias{overex2df}
\title{Obtain a data.frame from summarized experiment and group of genes using Fisher's test.}
\usage{
overex2df(sumex, hiperlink=FALSE)
}
\arguments{
\item{sumex}{Is the SummarizeExperiment object.}

\item{grpgenes}{Group of genes generated  getGenesets function.}

\item{hiperlink}{TRUE will include the corresponding hyperlinks to KEGG.}
}
\value{
data.frame
}
\description{
Generates a dataframe with the gsRanking information of the SummarizeExperiment and the group of genes. 
The SummarizedExperiment must have in the colData a column that is called 'GROUP' and must have '0' and '1' values.
Must have ENTRZID identifiers.
}
\seealso{
Other dataframe generation: 
\code{\link{sumex2dfDESeq2}()},
\code{\link{sumex2df}()}
}
\concept{dataframe generation}
