% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA548871}
\alias{PRJNA548871}
\title{PRJNA548871}
\format{
SummarizedExperiment
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE132731}
 
 \url{https://www.ncbi.nlm.nih.gov/Traces/study/?acc=PRJNA548871}
}
\description{
Mus musculus

SummarizedExperiment data from RNA-seq experiment in Mus musculus retinal microglia tissue.

Temporo-spatial distribution and transcriptional profile of retinal microglia in the oxygen-induced retinopathy mouse model.

RNA-Seq revealed an enrichment of processes related to cell division and chemokine receptor binding. We propose that activated
}
\examples{
data(PRJNA548871, package="asierortega")
}
\keyword{datasets}
