% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA548871_DE_analysis_edgeR}
\alias{PRJNA548871_DE_analysis_edgeR}
\title{PRJNA548871_DE_analysis_edgeR}
\format{
analysis
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE132731}
}
\description{
Mus musculus

Expression profiling by high throughput sequencing

SummarizedExperiment data from RNA-seq experiment in Mus musculus retinal microglia tissue.

Differential expression edgeR classic analysis with the data generated in PRJNA548871, tarea41.Rmd.'
}
\keyword{datasets}
