% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumex2df}
\alias{sumex2df}
\title{Obtain a data.frame from summarized experiment with p.values using edgeR.}
\usage{
sumex2df(sumex, columnName="treatment", method="c", hiperlink=FALSE)
}
\arguments{
\item{sumex}{Is the SummarizeExperiment object}

\item{columnName}{The name of the column with the information of the groups (treatment and control).
Default value will be "treatment".}

\item{method}{is the function that will use to obtain the p-values. 
"c" to apply the function of Common Dispersion.
"t" to apply the function of Tagwise Dispersion. 
As default value will apply the Common Dispersion function.}

\item{hiperlink}{TRUE will include the corresponding hyperlinks to the ENSEMBL and ENTREZID identifiers. 
FALSE to just include the identifiers without the corresponding hyperlinks.}
}
\value{
data.frame
}
\description{
Generates a dataframe with the rowData information of the SummarizeExperiment
and adds the p-value and adjusted p-value information using edgeR.
The column names of the rowData in the Summarized Experiment must contain the information
at least of ENSEMBL, GENENAME, SYMBOL and ENTREZID.
}
\seealso{
Other dataframe generation: 
\code{\link{overex2df}()},
\code{\link{sumex2dfDESeq2}()}
}
\concept{dataframe generation}
