% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unifyeset}
\alias{unifyeset}
\title{Unify the probes thar correspond to the same gene.}
\usage{
unifyeset(eset, ID = ENTREZID)
}
\arguments{
\item{eset}{is the ExpressionSet}

\item{ID}{Is the identifier that the probe have been annotated.}
}
\value{
ExpressionSet
}
\description{
Unify the probes thar correspond to the same gene.
It unify all the probes to correspond to the same gene identifier that is 
introduced by argument, as default value will be ENTREZID.
}
\seealso{
Other NA clean: 
\code{\link{cleanNA}()}
}
\concept{NA clean}
